refactor(api): extract deferred analytics routes from api.mjs
Move trajectory, uptime, leaderboards, and compare handlers into
workers/request-handlers/analytics-routes.mjs, fix missing await on
envelopeResponse before D1-fallback tagging, and add direct unit tests.

Closes #1917
