+++ workers/api.mjs
@@ -35,8 +35,6 @@ import {
   handleBadgeSvgRequest,
 } from "./request-handlers/discovery.mjs";
 import {
-  analyticsMeta,
-  analyticsQueryError,
   configureAnalytics,
   d1All,
   handleBulkHealthTrends,
@@ -44,9 +42,6 @@ import {
   handleHealthIncidents,
   handleHealthPercentiles,
   handleHealthTrends,
-  hasD1FallbackRows,
-  markD1FallbackResponse,
-  validateQueryParams,
   withEdgeCache,
 } from "./request-handlers/analytics.mjs";
 import {
@@ -76,6 +71,14 @@ import {
   handleExtrinsics,
   handleExtrinsic,
 } from "./request-handlers/entities.mjs";
+import {
+  canonicalCompareCachePath,
+  configureAnalyticsRoutes,
+  handleCompare,
+  handleLeaderboards,
+  handleTrajectory,
+  handleUptime,
+} from "./request-handlers/analytics-routes.mjs";
 import {
   classifyUpstreamAttempt,
   configureRpcProxy,
@@ -120,13 +123,8 @@ import {
   writeSubnetSnapshot,
 } from "../src/health-prober.mjs";
 import { KV_ECONOMICS_CURRENT } from "../src/kv-keys.mjs";
-import { dailyLatencyColumns } from "../src/health-sql.mjs";
 import {
   buildGlobalHealth,
-  formatLeaderboards,
-  formatTrajectory,
-  formatUptime,
-  LEADERBOARD_BOARDS,
   mergeFreshness,
   mergeRpcEndpoints,
   overlayArtifactEndpoints,
@@ -194,7 +192,6 @@ import {
   EXTRINSIC_DETAIL_PATH_PATTERN,
   EXTRINSICS_FEED_PATH_PATTERN,
   BULK_TRENDS_PATH_PATTERN,
-  DAY_MS,
   EMBEDDING_SYNC_CRON,
   EVENTS_INGEST_TOKEN_HEADER,
   EVENTS_LOAD_CRON,
@@ -208,7 +205,6 @@ import {
   MAX_BLOCKS_INGEST_ROWS,
   MAX_EVENTS_INGEST_BODY_BYTES,
   MAX_EVENTS_INGEST_ROWS,
-  MAX_UPTIME_ROWS,
   MAX_WEBHOOK_BODY_BYTES,
   NEURON_HISTORY_ROLLUP_CRON,
   PERCENTILES_PATH_PATTERN,
@@ -223,7 +219,6 @@ import {
   TRAJECTORY_PATH_PATTERN,
   TRENDS_PATH_PATTERN,
   UPTIME_PATH_PATTERN,
-  UPTIME_WINDOWS,
   WEBHOOK_SUBSCRIPTION_TOKEN_HEADER,
   WEBHOOK_TTL_SECONDS,
 } from "./config.mjs";
@@ -358,7 +353,9 @@ export {
   weightedPickEndpoint,
 };
 
-// Byte length of a UTF-8 string. Shared by the realtime ingest handlers below to
+export { composeCompareData } from "./request-handlers/analytics-routes.mjs";
+
+// Byte length of a UTF-8 string.
 // bound request bodies before parsing. (The staging loaders carry their own copy;
 // it is a pure stdlib one-liner, so a tiny duplicate beats a cross-module import
 // for a leaf used on both sides of the extraction.)
@@ -1678,12 +1675,7 @@ const SUBNET_SLUG_INDEX_TTL_MS = 300_000;
 // chains — testnet SN-N is unrelated to mainnet SN-N).
 const subnetSlugIndexByNetwork = new Map(); // network.id -> { map, builtAt }
 
-// Leaderboards re-derives a {meta, completeness} projection from the ~600 KB
-// R2 profiles.json; cache the small projection in-isolate (5 min TTL, same as
-// the slug index) so junk-query-param cache-busting can't force a full R2 read
-// + parse per request.
-const LEADERBOARD_PROFILES_TTL_MS = 300_000;
-let leaderboardProfilesCache = null; // { subnetMeta, mostComplete, builtAt }
+// Leaderboards/compare profiles projection cache lives in analytics-routes.mjs.
 
 // KV_HEALTH_META is written by the health cron (~15 min cadence) and read by
 // every analytics handler (percentiles, incidents, trends, uptime, trajectory,
@@ -1750,6 +1742,8 @@ export async function readEconomicsCurrentKv(env, now = Date.now()) {
   }
   return value;
 }
+
+configureAnalyticsRoutes({ readHealthMetaKv, readEconomicsCurrentKv });
 
 async function resolveSubnetSlugRoute(
   env,
@@ -2114,485 +2108,6 @@ async function handleApiRequest(
     ctx?.waitUntil?.(overlayCache.put(overlayCacheKey, response.clone()));
   }
   return response;
-}
-
-// Week-over-week structural trajectory from daily snapshots.
-async function handleTrajectory(request, env, netuid, url) {
-  const validationError = validateQueryParams(url, []);
-  if (validationError) return analyticsQueryError(validationError);
-  const rows = await d1All(
-    env,
-    `SELECT snapshot_date, completeness_score, surface_count, endpoint_count,
-            validator_count, miner_count, total_stake_tao, alpha_price_tao,
-            emission_share
-     FROM subnet_snapshots
-     WHERE netuid = ?
-     ORDER BY snapshot_date DESC
-     LIMIT 400`,
-    [netuid],
-  );
-  const data = formatTrajectory({ netuid, rows });
-  const response = envelopeResponse(
-    request,
-    {
-      data,
-      meta: await analyticsMeta(
-        env,
-        `/metagraph/subnets/${netuid}/trajectory.json`,
-        null,
-      ),
-    },
-    "short",
-  );
-  return hasD1FallbackRows(rows) ? markD1FallbackResponse(response) : response;
-}
-
-// Long-term daily uptime history for one subnet's operational surfaces, served
-// live from the surface_uptime_daily rollup (PR3). 90d/1y window. Returns a
-// schema-stable empty payload when D1 is unbound/cold or no history has accrued
-// yet (mirrors the other D1-backed analytics routes).
-async function handleUptime(request, env, netuid, url) {
-  const validationError = validateQueryParams(url, ["window"]);
-  if (validationError) return analyticsQueryError(validationError);
-  const windowParam = url.searchParams.get("window") || "90d";
-  if (!Object.hasOwn(UPTIME_WINDOWS, windowParam)) {
-    return errorResponse(
-      "invalid_query",
-      "Query parameter `window` must be one of: 90d, 1y.",
-      400,
-      { parameter: "window" },
-    );
-  }
-  const days = UPTIME_WINDOWS[windowParam];
-  const cutoff = new Date(Date.now() - days * 24 * 60 * 60 * 1000)
-    .toISOString()
-    .slice(0, 10);
-  const rows = await d1All(
-    env,
-    `SELECT MAX(surface_id) AS surface_id,
-            COALESCE(surface_key, surface_id) AS surface_key,
-            day,
-            SUM(samples) AS samples,
-            SUM(ok_count) AS ok_count,
-            CASE
-              WHEN SUM(samples) > 0 THEN ROUND(CAST(SUM(ok_count) AS REAL) / SUM(samples), 4)
-              ELSE NULL
-            END AS uptime_ratio,
-            ${dailyLatencyColumns({ roundedAvg: true })},
-            MAX(p50_latency_ms) AS p50,
-            MAX(p95_latency_ms) AS p95,
-            MAX(p99_latency_ms) AS p99,
-            CASE
-              WHEN SUM(samples) = 0 THEN 'unknown'
-              WHEN SUM(ok_count) = SUM(samples) THEN 'ok'
-              WHEN SUM(ok_count) = 0 THEN 'failed'
-              ELSE 'degraded'
-            END AS status
-     FROM surface_uptime_daily
-     WHERE netuid = ? AND day >= ?
-     GROUP BY COALESCE(surface_key, surface_id), day
-     ORDER BY day DESC
-     LIMIT ?`,
-    [netuid, cutoff, MAX_UPTIME_ROWS],
-  );
-  const healthMeta = await readHealthMetaKv(env);
-  const data = formatUptime({
-    netuid,
-    window: windowParam,
-    observedAt: healthMeta?.last_run_at || null,
-    rows,
-    now: new Date().toISOString(),
-  });
-  const response = envelopeResponse(
-    request,
-    {
-      data,
-      meta: await analyticsMeta(
-        env,
-        `/metagraph/subnets/${netuid}/uptime.json`,
-        data.observed_at,
-      ),
-    },
-    "short",
-  );
-  return hasD1FallbackRows(rows) ? markD1FallbackResponse(response) : response;
-}
-
-// Small {meta, completeness} projection over profiles.json, cached in-isolate.
-async function leaderboardProfilesProjection(env, now = Date.now()) {
-  if (
-    leaderboardProfilesCache &&
-    now - leaderboardProfilesCache.builtAt <= LEADERBOARD_PROFILES_TTL_MS
-  ) {
-    return leaderboardProfilesCache;
-  }
-  const artifact = await readArtifact(env, "/metagraph/profiles.json");
-  const profiles = artifact.ok ? artifact.data?.profiles || [] : [];
-  const subnetMeta = new Map();
-  const mostComplete = [];
-  for (const profile of profiles) {
-    if (!Number.isInteger(profile.netuid)) continue;
-    subnetMeta.set(profile.netuid, {
-      slug: profile.slug ?? null,
-      name: profile.name ?? null,
-    });
-    mostComplete.push({
-      netuid: profile.netuid,
-      slug: profile.slug ?? null,
-      name: profile.name ?? null,
-      completeness_score: profile.completeness_score ?? null,
-      // Enrichment-depth signals for the most-enriched board (#753).
-      surface_count: profile.surface_count ?? 0,
-      operational_interface_count: profile.operational_interface_count ?? 0,
-    });
-  }
-  const projection = { subnetMeta, mostComplete, builtAt: now };
-  // Don't cache an empty projection (failed/cold read) — retry next request.
-  if (mostComplete.length > 0) {
-    leaderboardProfilesCache = projection;
-  }
-  return projection;
-}
-
-// Economic source for cross-subnet projections (e.g. the leaderboard economic
-// boards): the live KV 'economics:current' blob when fresh/on-contract/integrity-
-// checked, else the committed R2 economics.json — the same KV-primary/R2-fallback
-// the /api/v1/economics route uses. Always resolves to an array (empty when both
-// tiers are cold), so every caller stays null-safe.
-async function resolveEconomicsRows(env) {
-  const live = await resolveLiveEconomics({
-    readHealthKv,
-    env,
-    contractVersion: contractVersion(env),
-  });
-  if (Array.isArray(live?.data?.subnets)) return live.data.subnets;
-  const artifact = await readArtifact(env, "/metagraph/economics.json");
-  return artifact.ok && Array.isArray(artifact.data?.subnets)
-    ? artifact.data.subnets
-    : [];
-}
-
-// Registry leaderboards: healthiest / fastest-rpc / most-complete /
-// most-enriched / fastest-growing plus the economic opportunity boards
-// (open-slots / cheapest-registration / highest-emission / validator-headroom).
-// Combines live D1 status with registry projections and the economics tier.
-async function handleLeaderboards(request, env, url) {
-  const validationError = validateQueryParams(url, ["board", "limit"]);
-  if (validationError) return analyticsQueryError(validationError);
-  const requestedBoard = url.searchParams.get("board");
-  if (requestedBoard && !LEADERBOARD_BOARDS.includes(requestedBoard)) {
-    return errorResponse(
-      "invalid_query",
-      `Unknown board "${requestedBoard}". Valid boards: ${LEADERBOARD_BOARDS.join(", ")}.`,
-      400,
-    );
-  }
-  const limit = url.searchParams.get("limit");
-  if (
-    limit !== null &&
-    (!/^\d+$/.test(limit) || Number(limit) < 1 || Number(limit) > 100)
-  ) {
-    // Reject invalid limits with a 400 like the list routes, instead of the
-    // silent clamp formatLeaderboards would otherwise apply.
-    return errorResponse(
-      "invalid_query",
-      "limit must be an integer between 1 and 100.",
-      400,
-    );
-  }
-
-  const { subnetMeta, mostComplete } = await leaderboardProfilesProjection(env);
-
-  const sevenDaysAgo = new Date(Date.now() - 7 * DAY_MS)
-    .toISOString()
-    .slice(0, 10);
-  const [healthRows, rpcRows, growthSamples, economicsRows] = await Promise.all(
-    [
-      d1All(
-        env,
-        `SELECT netuid,
-              COUNT(*) AS total,
-              SUM(CASE WHEN status = 'ok' THEN 1 ELSE 0 END) AS ok_count,
-              AVG(latency_ms) AS avg_latency_ms
-       FROM surface_status
-       GROUP BY netuid`,
-        [],
-      ),
-      d1All(
-        env,
-        `SELECT netuid, MIN(latency_ms) AS min_latency_ms
-       FROM surface_status
-       WHERE kind IN ('subtensor-rpc', 'subtensor-wss')
-         AND status = 'ok' AND latency_ms IS NOT NULL
-       GROUP BY netuid`,
-        [],
-      ),
-      d1All(
-        env,
-        `SELECT netuid, snapshot_date, completeness_score
-       FROM subnet_snapshots
-       WHERE snapshot_date >= ?
-       ORDER BY netuid, snapshot_date`,
-        [sevenDaysAgo],
-      ),
-      // Economic boards: the economics tier alongside the operational D1 queries.
-      resolveEconomicsRows(env),
-    ],
-  );
-
-  // Per-subnet completeness delta over the window (latest - earliest sample).
-  const growthByNetuid = new Map();
-  for (const row of growthSamples) {
-    const entry = growthByNetuid.get(row.netuid) || {
-      first: undefined,
-      last: undefined,
-    };
-    // `undefined` = no row yet; a real null completeness_score must latch as the
-    // baseline so the delta guard below can drop unscored window endpoints.
-    if (entry.first === undefined) entry.first = row.completeness_score ?? null;
-    entry.last = row.completeness_score ?? null;
-    growthByNetuid.set(row.netuid, entry);
-  }
-  const growthRows = [...growthByNetuid.entries()].map(([netuid, entry]) => ({
-    netuid,
-    delta:
-      entry.first != null && entry.last != null
-        ? Number(entry.last) - Number(entry.first)
-        : null,
-  }));
-
-  const meta = await readHealthMetaKv(env);
-  const data = formatLeaderboards({
-    board: requestedBoard || null,
-    limit,
-    observedAt: meta?.last_run_at || null,
-    healthRows,
-    rpcRows,
-    mostComplete,
-    growthRows,
-    economicsRows,
-    subnetMeta,
-  });
-  const response = envelopeResponse(
-    request,
-    {
-      data,
-      meta: {
-        artifact_path: "/metagraph/registry/leaderboards.json",
-        cache: "standard",
-        contract_version: contractVersion(env),
-        generated_at: data.observed_at,
-        source: "registry+live-cron-prober",
-      },
-    },
-    "standard",
-  );
-  return hasD1FallbackRows(healthRows, rpcRows, growthSamples)
-    ? markD1FallbackResponse(response)
-    : response;
-}
-
-// The data domains /api/v1/compare can place side by side: registry structure
-// (completeness + surface counts from profiles), the live economics tier, and
-// the live per-subnet probe-health rollup. Composed in one call so a caller can
-// choose between subnets without N×(detail + economics + health) round-trips.
-const COMPARE_DIMENSIONS = ["structure", "economics", "health"];
-// Same shape + hard cap as the subnets collection's `netuids` CSV filter: 1-128
-// ids, each ≤ 5 digits. Bounds the compare fan-out at the parameter layer.
-const COMPARE_NETUIDS_PATTERN = /^\d{1,5}(,\d{1,5}){0,127}$/;
-
-function compareNetuids(netuidsRaw) {
-  if (!netuidsRaw || !COMPARE_NETUIDS_PATTERN.test(netuidsRaw)) return null;
-  const requestedNetuids = [];
-  const seenNetuids = new Set();
-  for (const part of netuidsRaw.split(",")) {
-    const netuid = Number(part);
-    if (seenNetuids.has(netuid)) continue;
-    seenNetuids.add(netuid);
-    requestedNetuids.push(netuid);
-  }
-  return requestedNetuids;
-}
-
-function compareDimensions(dimensionsRaw) {
-  if (dimensionsRaw === null) return COMPARE_DIMENSIONS;
-  const requested = dimensionsRaw.split(",");
-  const unknown = requested.find((d) => !COMPARE_DIMENSIONS.includes(d));
-  if (unknown !== undefined) return null;
-  return COMPARE_DIMENSIONS.filter((d) => requested.includes(d));
-}
-
-function canonicalCompareCachePath(url) {
-  if (validateQueryParams(url, ["netuids", "dimensions"])) return null;
-  const requestedNetuids = compareNetuids(url.searchParams.get("netuids"));
-  if (!requestedNetuids) return null;
-  const dimensions = compareDimensions(url.searchParams.get("dimensions"));
-  if (!dimensions) return null;
-  const params = [`netuids=${encodeURIComponent(requestedNetuids.join(","))}`];
-  if (dimensions.length !== COMPARE_DIMENSIONS.length) {
-    params.push(`dimensions=${encodeURIComponent(dimensions.join(","))}`);
-  }
-  return `${url.pathname}?${params.join("&")}`;
-}
-
-// Pure projection: fold the requested netuids + the resolved source rows into
-// the side-by-side compare shape, in REQUESTED order. A netuid absent from the
-// registry profiles is returned `found: false` with every requested dimension
-// null (so a caller can still align columns); a found subnet missing from a
-// given source tier gets that one dimension as null. Exported for unit coverage.
-export function composeCompareData({
-  requestedNetuids,
-  dimensions,
-  subnetMeta,
-  structureRows,
-  economicsRows,
-  healthRows,
-  observedAt,
-}) {
-  const includeStructure = dimensions.includes("structure");
-  const includeEconomics = dimensions.includes("economics");
-  const includeHealth = dimensions.includes("health");
-
-  const structureByNetuid = new Map();
-  for (const row of structureRows || []) {
-    structureByNetuid.set(row.netuid, {
-      completeness_score: row.completeness_score,
-      surface_count: row.surface_count,
-      operational_interface_count: row.operational_interface_count,
-    });
-  }
-  const economicsByNetuid = new Map();
-  for (const row of economicsRows || []) {
-    economicsByNetuid.set(row.netuid, {
-      registration_cost_tao: row.registration_cost_tao,
-      registration_allowed: row.registration_allowed,
-      open_slots: row.open_slots,
-      emission_share: row.emission_share,
-      alpha_price_tao: row.alpha_price_tao,
-      validator_count: row.validator_count,
-      miner_count: row.miner_count,
-      total_stake_tao: row.total_stake_tao,
-      miner_readiness: row.miner_readiness,
-    });
-  }
-  const healthByNetuid = new Map();
-  for (const row of healthRows || []) {
-    healthByNetuid.set(row.netuid, {
-      surface_count: row.surface_count,
-      ok_count: row.ok_count,
-      avg_latency_ms: row.avg_latency_ms,
-    });
-  }
-
-  const subnets = requestedNetuids.map((netuid) => {
-    const meta = subnetMeta.get(netuid) || null;
-    const entry = {
-      netuid,
-      name: meta?.name ?? null,
-      slug: meta?.slug ?? null,
-      found: meta !== null,
-    };
-    if (includeStructure) {
-      entry.structure = meta ? (structureByNetuid.get(netuid) ?? null) : null;
-    }
-    if (includeEconomics) {
-      entry.economics = meta ? (economicsByNetuid.get(netuid) ?? null) : null;
-    }
-    if (includeHealth) {
-      entry.health = meta ? (healthByNetuid.get(netuid) ?? null) : null;
-    }
-    return entry;
-  });
-
-  return {
-    schema_version: 1,
-    source: "registry+economics+live-cron-prober",
-    observed_at: observedAt ?? null,
-    dimensions,
-    requested_netuids: requestedNetuids,
-    subnets,
-  };
-}
-
-// Cross-subnet compare: place several subnets side by side across the registry
-// structure, economics, and live-health tiers in one call (fileless-D1 pattern,
-// like handleLeaderboards). `netuids` is required; `dimensions` defaults to all.
-async function handleCompare(request, env, url) {
-  const validationError = validateQueryParams(url, ["netuids", "dimensions"]);
-  if (validationError) return analyticsQueryError(validationError);
-
-  const netuidsRaw = url.searchParams.get("netuids");
-  const requestedNetuids = compareNetuids(netuidsRaw);
-  if (!requestedNetuids) {
-    return errorResponse(
-      "invalid_query",
-      "netuids is required: a comma-separated list of 1-128 subnet ids.",
-      400,
-      { parameter: "netuids" },
-    );
-  }
-
-  const dimensionsRaw = url.searchParams.get("dimensions");
-  const dimensions = compareDimensions(dimensionsRaw);
-  if (!dimensions) {
-    const unknown = dimensionsRaw
-      .split(",")
-      .find((d) => !COMPARE_DIMENSIONS.includes(d));
-    return errorResponse(
-      "invalid_query",
-      `Unknown dimension "${unknown}". Valid dimensions: ${COMPARE_DIMENSIONS.join(", ")}.`,
-      400,
-      { parameter: "dimensions" },
-    );
-  }
-
-  // subnetMeta + structure always come from the cached profiles projection;
-  // economics + health are only read when their dimension is requested.
-  const { subnetMeta, mostComplete } = await leaderboardProfilesProjection(env);
-  const [economicsRows, healthRows] = await Promise.all([
-    dimensions.includes("economics") ? resolveEconomicsRows(env) : null,
-    dimensions.includes("health")
-      ? d1All(
-          env,
-          `SELECT netuid,
-                COUNT(*) AS surface_count,
-                SUM(CASE WHEN status = 'ok' THEN 1 ELSE 0 END) AS ok_count,
-                ROUND(AVG(latency_ms)) AS avg_latency_ms
-         FROM surface_status
-         WHERE netuid IN (${requestedNetuids.map(() => "?").join(", ")})
-         GROUP BY netuid`,
-          requestedNetuids,
-        )
-      : null,
-  ]);
-
-  const meta = await readHealthMetaKv(env);
-  const data = composeCompareData({
-    requestedNetuids,
-    dimensions,
-    subnetMeta,
-    structureRows: mostComplete,
-    economicsRows,
-    healthRows,
-    observedAt: meta?.last_run_at ?? null,
-  });
-  const response = envelopeResponse(
-    request,
-    {
-      data,
-      meta: {
-        artifact_path: "/metagraph/compare.json",
-        cache: "standard",
-        contract_version: contractVersion(env),
-        generated_at: data.observed_at,
-        source: "registry+economics+live-cron-prober",
-      },
-    },
-    "standard",
-  );
-  return hasD1FallbackRows(healthRows)
-    ? markD1FallbackResponse(response)
-    : response;
 }
 
 function matchRawArtifact(pathname) {
